jwt_generator/requirements.txt
PyJWT
fastapi==0.115.5
pydantic-settings==2.6.1
httpx